Switch to source from pypi
This solves an issue where the version in the source from
GitHub indicates that the version number is `4.2.dev0`, which gives
us trouble with downstream package when we run `pip check`.

This merge also adds `pip check` to the tests.

This merge removes C compilers and cython because these don't make
sense for a `noarch:python` package.  If `cython` is needed, this
package need to be compiled for each python version and OS.

Finally, this merge adds me (@xylar) as a maintainer.